requirements.txt
Django==4.1.9
mssql-django==1.2
django-cors-headers==4.0.0
pyodbc==4.0.39
Unidecode==1.3.6
sqlparse==0.4.4
djangorestframework==3.15.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
mysqlclient
python-decouple
gunicorn